requirements.txt
amqp==2.6.1
asgiref==3.5.2
billiard==3.6.4.0
boto3==1.26.3
botocore==1.29.3
celery==4.4.1
crispy-tailwind==0.5.0
debugpy==1.6.3
Django==4.1
django-appconf==1.0.5
django-browser-reload==1.6.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-extensions==3.2.1
django-htmx==1.12.2
django-imagekit==4.1.0
django-render-block==0.9.1
django-storages==1.13.1
django-widget-tweaks==1.4.12
jmespath==1.0.1
kombu==4.6.11
pilkit==2.0
Pillow==9.2.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.4
redis==3.4.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
vine==1.3.0
